### 2021-11-17 #486 @wlandau
gittargets: data version control for the targets package
Submitting Author Name: Will Landau
Submitting Author Github Handle: @wlandau
Repository:  https://github.com/wlandau/gittargets
Version submitted: 0.0.0.9000
Submission type: Standard
Version accepted: TBD
---

```
Package: gittargets
Title: Data Version Control for the Targets Package
Description: Version control systems such as Git help researchers
  track changes and history in data science projects,
  and the 'targets' package (2021, <doi:10.21105/joss.02959>)
  minimizes the computational cost of keeping the latest results
  reproducible and up to date. The 'gittargets' package
  combines these two capabilities. The 'targets' data store
  becomes a version control repository and stays synchronized
  with the Git repository of the source code.
  Users can switch commits and branches
  without invalidating the 'targets' pipeline.
Version: 0.0.0.9000
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ))
  cli (>= 3.0.0),
  data.table (>= 1.12.8),
  gert (>= 1.0.0),
  processx (>= 3.0.0),
  targets (>= 0.8.1.9000),
  tibble (>= 3.0.0),
  uuid (>= 1.0.0)
  knitr (>= 1.30),
  markdown (>= 1.1),
  rmarkdown (>= 2.4),
  ropensci/targets
SystemRequirements: Git (>= 2.0.0)
```

Version control systems such as Git help researchers track changes and history in data science projects, and the targets package minimizes the computational cost of keeping the latest results reproducible and up to date. The gittargets package combines these two capabilities. The targets data store becomes a version control repository and stays synchronized with the Git repository of the source code. Users can switch commits and branches without invalidating the targets pipeline.

`gittargets` is for people who use [`targets`](https://docs.ropensci.org/targets/) for reproducible analysis pipelines and Git for tracking the code files of those pipelines.

Packages `gh`, `gert`, `git2r`, and `git2rdata` all work with Git from R, but in a way that is more general, less opinionated, and less [`targets`](https://docs.ropensci.org/targets/)-specific than `gittargets`. `gittargets` establishes a separate repository for the [`targets`](https://docs.ropensci.org/targets/) data that is linked to the project's code repository. This allows the code and data to synchronize without affecting the efficiency or the size of the code repository.

N/A

N/A

- [n/a] The package is novel and will be of interest to the broad readership of the journal.
- [n/a] The manuscript describing the package is no longer than 3000 words.
- [n/a] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal (see [MEE's Policy on Publishing Code](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
