### 2021-11-17 1c64c3f
Use message stream for printing data frames
