+++ tests/testthat/test-utils_assert.R
@@ -5,7 +5,7 @@ targets::tar_test("tar_git_assert_commits_code()", {
     tar_git_assert_commits_code(getwd()),
     class = "tar_condition_validate"
   )
-  utils::capture.output(git_setup_init())
+  git_setup_init()
   expect_silent(tar_git_assert_commits_code(getwd()))
 })
 
@@ -16,7 +16,7 @@ targets::tar_test("tar_git_assert_commits_data()", {
     tar_git_assert_commits_data(getwd()),
     class = "tar_condition_validate"
   )
-  utils::capture.output(git_setup_init())
+  git_setup_init()
   expect_silent(tar_git_assert_commits_data(getwd()))
 })
 
@@ -42,13 +42,13 @@ targets::tar_test("tar_git_assert_repo_data()", {
 
 targets::tar_test("tar_git_assert_repo_data() no branch", {
   skip_os_git()
-  utils::capture.output(git_setup_init())
+  git_setup_init()
   store <- targets::tar_config_get("store")
   expect_error(
     tar_git_assert_snapshot(branch = "nope", store = store),
     class = "tar_condition_validate"
   )
-  utils::capture.output(tar_git_snapshot(status = FALSE, verbose = FALSE))
+  tar_git_snapshot(status = FALSE, verbose = FALSE)
   branch <- tar_git_branch_snapshot(tar_git_log()$commit_code)
   expect_silent(tar_git_assert_snapshot(branch = branch, store = store))
 })
