+++ R/tar_package.R
@@ -12,6 +12,7 @@
 #' @name gittargets-package
 #' @family help
 #' @importFrom cli cli_alert_danger cli_alert_info cli_alert_success cli_h1
+#'   col_br_white
 #' @importFrom data.table as.data.table
 #' @importFrom gert git_branch git_branch_exists git_branch_list
 #'   git_commit_info git_log git_status
@@ -20,6 +21,6 @@
 #' @importFrom targets tar_assert_file tar_config_get tar_config_set tar_dir
 #'   tar_outdated tar_throw_validate
 #' @importFrom tibble as_tibble tibble
-#' @importFrom utils head menu
+#' @importFrom utils capture.output head menu
 #' @importFrom uuid UUIDgenerate
 NULL
