+++ tests/testthat/test-tar_git_checkout.R
@@ -9,9 +9,9 @@ targets::tar_test("tar_git_checkout()", {
   gert::git_add("_targets.R")
   gert::git_commit("First commit")
   gert::git_branch_create("old_branch")
-  utils::capture.output(tar_git_init())
+  tar_git_init()
   # Work on a new branch.
-  utils::capture.output(tar_git_snapshot(status = FALSE, verbose = FALSE))
+  tar_git_snapshot(status = FALSE, verbose = FALSE)
   targets::tar_script(tar_target(data, "new_data"))
   targets::tar_make(callr_function = NULL)
   expect_equal(targets::tar_progress(data)$progress, "built")
@@ -19,13 +19,13 @@ targets::tar_test("tar_git_checkout()", {
   gert::git_branch_create("new_branch")
   gert::git_add("_targets.R")
   gert::git_commit("Second commit")
-  utils::capture.output(tar_git_snapshot(status = FALSE, verbose = FALSE))
+  tar_git_snapshot(status = FALSE, verbose = FALSE)
   # Go back to the old branch.
   gert::git_branch_checkout("old_branch")
   # The target is out of date because we only reverted the code.
   targets::tar_outdated(callr_function = NULL)
   # But tar_git_checkout() lets us restore the old version of the data!
-  utils::capture.output(tar_git_checkout())
+  tar_git_checkout()
   # Now, the target is up to date! And we did not even have to rerun it!
   expect_equal(targets::tar_outdated(callr_function = NULL), character(0))
   expect_equal(targets::tar_read(data), "old_data")
