+++ R/utils_git.R
@@ -1,24 +1,54 @@
-tar_git_add <- function(files, repo, echo = TRUE, spinner = TRUE) {
+tar_git_add <- function(files, repo, spinner = TRUE) {
   processx::run(
     command = tar_git_binary(),
     args = c("add", files),
     wd = repo,
-    echo = echo,
+    echo = FALSE,
     spinner = spinner
   )
 }
 
 tar_git_branch_checkout <- function(branch, repo, force) {
   args <- c("checkout", if_any(force, "--force", character(0)), branch)
-  processx::run(command = tar_git_binary(), args = args, wd = repo)
+  processx::run(
+    command = tar_git_binary(),
+    args = args,
+    wd = repo,
+    echo = FALSE
+  )
 }
 
 tar_git_branch_create <- function(branch, repo) {
   processx::run(
     command = tar_git_binary(),
     args = c("branch", branch),
-    wd = repo
+    wd = repo,
+    echo = FALSE
   )
+}
+
+tar_git_commit <- function(message, repo, spinner = TRUE) {
+  processx::run(
+    command = tar_git_binary(),
+    args = c("commit", "--message", message),
+    wd = repo,
+    echo = FALSE,
+    spinner = spinner
+  )
+}
+
+tar_git_commit_all <- function(message, repo, spinner = TRUE) {
+  processx::run(
+    command = tar_git_binary(),
+    args = c("commit", "--all", "--message", message),
+    wd = repo,
+    echo = FALSE,
+    spinner = spinner
+  )
+}
+
+tar_git_init_repo <- function(path) {
+  processx::run(command = tar_git_binary(), args = "init", wd = path)
 }
 
 tar_git_branch_snapshot <- function(commit) {
@@ -27,30 +57,6 @@ tar_git_branch_snapshot <- function(commit) {
 
 tar_git_commit_code <- function(branch) {
   gsub(pattern = "^code=", replacement = "", x = branch)
-}
-
-tar_git_commit <- function(message, repo, echo = TRUE, spinner = TRUE) {
-  processx::run(
-    command = tar_git_binary(),
-    args = c("commit", "--message", message),
-    wd = repo,
-    echo = echo,
-    spinner = spinner
-  )
-}
-
-tar_git_commit_all <- function(message, repo, echo = TRUE, spinner = TRUE) {
-  processx::run(
-    command = tar_git_binary(),
-    args = c("commit", "--all", "--message", message),
-    wd = repo,
-    echo = echo,
-    spinner = spinner
-  )
-}
-
-tar_git_init_repo <- function(path) {
-  processx::run(command = tar_git_binary(), args = "init", wd = path)
 }
 
 tar_git_binary <- function() {
