### 2021-11-17 ca36b8d
Rename functions
