### 2021-11-17 #486 @wlandau
`gittargets` requires command line Git in order for `git-lfs` to work (related: https://github.com/r-lib/gert/issues/160). The majority of tests will automatically be skipped if `tar_git_ok()` returns `FALSE`. Is there still a path forward?
