### 2021-11-17 78f8461
Test more Git utils
