### 2021-11-17 #486 @wlandau
Thank you both!

Related: I have not been able to correctly set up Git on the Windows GitHub Actions for `gittargets`. I tried a bunch of different approaches, but I still get errors claiming I need to configure `git config --global user.name` and `git config --global user.email`.
