+++ .github/workflows/cover.yaml
@@ -7,10 +7,6 @@ jobs:
     runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-      AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-      AWS_DEFAULT_REGION: ${{ secrets.AWS_DEFAULT_REGION }}
-      TORCH_INSTALL: 1
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
     steps:
       - uses: actions/checkout@v2
