+++ .github/workflows/cover-ubuntu.yaml
@@ -25,8 +25,8 @@ jobs:
         uses: actions/cache@v1
         with:
           path: ${{ env.R_LIBS_USER }}
-          key: macOS-r-4.0-2-${{ hashFiles('.github/depends.Rds') }}
-          restore-keys: macOS-r-4.0-2-
+          key: ubuntu-latest-1-${{ hashFiles('.github/depends.Rds') }}
+          restore-keys: ubuntu-latest-1-
 
       - name: Install packages
         run: |
