### 2021-11-18 6863557
Allow compat with targets 0.6.0
Need to clone a little assertion, but worth it for now.
