+++ tests/testthat/test-utils_assert.R
@@ -1,3 +1,13 @@
+# TODO: remove when it is safe to import
+# tar_assert_file() from `targets`.
+tar_test("tar_assert_file()", {
+  expect_error(tar_assert_file(0), class = "tar_condition_validate")
+  expect_error(tar_assert_file("x"), class = "tar_condition_validate")
+  file.create("x")
+  expect_error(tar_assert_file(c("x", "y")), class = "tar_condition_validate")
+  expect_silent(tar_assert_file("x"))
+})
+
 targets::tar_test("tar_git_assert_commits_code()", {
   skip_os_git()
   gert::git_init()
