+++ R/tar_git_status_code.R
@@ -19,7 +19,7 @@
 #' })
 #' }
 tar_git_status_code <- function(code = getwd()) {
-  targets::tar_assert_file(code)
+  tar_assert_file(code)
   if_any(
     tar_git_repo_exists(repo = code),
     gert::git_status(repo = code),
