+++ R/tar_package.R
@@ -18,7 +18,7 @@
 #'   git_commit_info git_log git_status
 #' @importFrom processx run
 #' @importFrom stats complete.cases
-#' @importFrom targets tar_assert_file tar_config_get tar_config_set tar_dir
+#' @importFrom targets tar_config_get tar_config_set tar_dir
 #'   tar_outdated tar_throw_validate
 #' @importFrom tibble as_tibble tibble
 #' @importFrom utils capture.output head menu
