+++ tests/testthat/test-tar_git_init.R
@@ -41,6 +41,7 @@ targets::tar_test("tar_git_init() with .gitignore", {
   store <- targets::tar_config_get("store")
   targets::tar_script(targets::tar_target(x, 1))
   targets::tar_make(callr_function = NULL)
+  writeLines("*", file.path(store, ".gitignore"))
   expect_false(file.exists(file.path(store, ".git")))
   expect_true(file.exists(file.path(store, ".gitignore")))
   dir <- getwd()
