### 2021-11-18 7d8cfc8
Namespace function
