+++ tests/testthat/test-utils_assert.R
@@ -1,6 +1,6 @@
 # TODO: remove when it is safe to import
 # tar_assert_file() from `targets`.
-tar_test("tar_assert_file()", {
+targets::tar_test("tar_assert_file()", {
   expect_error(tar_assert_file(0), class = "tar_condition_validate")
   expect_error(tar_assert_file("x"), class = "tar_condition_validate")
   file.create("x")
