+++ README.Rmd
@@ -13,10 +13,10 @@ knitr::opts_chunk$set(
 
 # gittargets <img src='man/figures/logo-readme.png' align="right" height="139"/>
 
-[![CRAN](https://www.r-pkg.org/badges/version/gittargets)](https://CRAN.R-project.org/package=gittargets)
+<!--[![CRAN](https://www.r-pkg.org/badges/version/gittargets)](https://CRAN.R-project.org/package=gittargets)-->
 [![status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![check](https://github.com/wlandau/gittargets/workflows/check/badge.svg)](https://github.com/wlandau/gittargets/actions?query=workflow%3Acheck)
-[![codecov](https://codecov.io/gh/wlandau/gittargets/branch/main/graph/badge.svg?token=3T5DlLwUVl)](https://codecov.io/gh/wlandau/gittargets)
+[![codecov](https://codecov.io/gh/wlandau/gittargets/branch/main/graph/badge.svg?token=3T5DlLwUVl)](https://app.codecov.io/gh/wlandau/gittargets)
 [![lint](https://github.com/wlandau/gittargets/workflows/lint/badge.svg)](https://github.com/wlandau/gittargets/actions?query=workflow%3Alint)
 
 Version control systems such as Git help researchers track changes and history in data science projects, and the [`targets`](https://docs.ropensci.org/targets/) package minimizes the computational cost of keeping the latest results reproducible and up to date. The `gittargets` package combines these two capabilities. The [`targets`](https://docs.ropensci.org/targets/) data store becomes a version control repository and stays synchronized with the Git repository of the source code. Users can switch commits and branches without invalidating the [`targets`](https://docs.ropensci.org/targets/) pipeline.
