### 2021-12-07 #486 @adamhsparks
@smwindecker and @mdneuzerling, thank you both for agreeing to review {gittargets} for us. I’m aware that this review and the three weeks may have competing priorities for your time with the holidays and leave at the end of the year/beginning of the new year. With that in mind, please just keep me posted on your progress as you have time. If you get these finished by sometime in January that would be fine, but we’re pretty flexible here and understand that life happens so please do ask if you need more time for any reason.
