### 2021-12-14 #486 @smwindecker
thanks @adamhsparks. My goal was to spend the next few days tinkering so will make a start in any case. I'm using targets for my main work project but still feel a bit junior league on it, particularly related to how best to incorporate the models that are sent to HPC into my targets workflow. Look forward to spending the time on this, but it may be first weeks of Jan.
