+++ .github/ISSUE_TEMPLATE/performance.md
@@ -9,8 +9,8 @@ assignees: wlandau
 
 ## Prework
 
-* [ ] Read and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html) and [contributing guidelines](https://github.com/wlandau/targit/blob/main/CONTRIBUTING.md).
-* [ ] If there is [already a relevant issue](https://github.com/wlandau/targit/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
+* [ ] Read and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html) and [contributing guidelines](https://github.com/wlandau/gittargets/blob/main/CONTRIBUTING.md).
+* [ ] If there is [already a relevant issue](https://github.com/wlandau/gittargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
 * [ ] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/wlandau/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
     * [ ] **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
     * [ ] **Minimal**: reduce runtime wherever possible and remove complicated details that are irrelevant to the issue at hand.
@@ -29,12 +29,12 @@ Please describe the performance issue.
 
 ## Benchmarks
 
-How poorly does `targit` perform? To find out, we recommend you use the [`proffer`](https://github.com/r-prof/proffer) package and take screenshots of the results displayed in your browser.
+How poorly does `gittargets` perform? To find out, we recommend you use the [`proffer`](https://github.com/r-prof/proffer) package and take screenshots of the results displayed in your browser.
 
 ```r
-library(targit)
+library(gittargets)
 library(proffer)
 px <- pprof({
-  # All your targit code goes here.
+  # All your gittargets code goes here.
 })
 ```
