+++ .github/ISSUE_TEMPLATE/bug.md
@@ -1,6 +1,6 @@
 ---
 name: Bug
-about: Please do not submit a bug report unless your issue is a genuine bug in targit and not a known limitation, usage error, or issue from another package that targit depends on.
+about: Please do not submit a bug report unless your issue is a genuine bug in gittargets and not a known limitation, usage error, or issue from another package that gittargets depends on.
 title: ""
 labels: "type: bug"
 assignees: wlandau
@@ -8,9 +8,9 @@ assignees: wlandau
 
 ## Prework
 
-* [ ] Read and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html) and [contributing guidelines](https://github.com/wlandau/targit/blob/main/CONTRIBUTING.md).
-* [ ] Confirm that your issue is a genuine bug in the `targit` package itself and not a user error, known limitation, or issue from another package that `targit` depends on. For example, if you get errors running `tar_make_clustermq()`, try isolating the problem in a reproducible example that runs `clustermq` and not `targit`. And for miscellaneous troubleshooting, please post to [discussions](https://github.com/wlandau/targit/discussions) instead of [issues](https://github.com/wlandau/targit/issues).
-* [ ] If there is [already a relevant issue](https://github.com/wlandau/targit/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
+* [ ] Read and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html) and [contributing guidelines](https://github.com/wlandau/gittargets/blob/main/CONTRIBUTING.md).
+* [ ] Confirm that your issue is a genuine bug in the `gittargets` package itself and not a user error, known limitation, or issue from another package that `gittargets` depends on. For example, if you get errors running `tar_make_clustermq()`, try isolating the problem in a reproducible example that runs `clustermq` and not `gittargets`. And for miscellaneous troubleshooting, please post to [discussions](https://github.com/wlandau/gittargets/discussions) instead of [issues](https://github.com/wlandau/gittargets/issues).
+* [ ] If there is [already a relevant issue](https://github.com/wlandau/gittargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
 * [ ] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/wlandau/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
     * [ ] **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
     * [ ] **Minimal**: reduce runtime wherever possible and remove complicated details that are irrelevant to the issue at hand.
@@ -36,4 +36,4 @@ What should have happened? Please be as specific as possible.
 * A [reproducible example](https://github.com/tidyverse/reprex).
 * Session info, available through `sessionInfo()` or [`reprex(si = TRUE)`](https://github.com/tidyverse/reprex).
 * A stack trace from `traceback()` or `rlang::trace_back()`.
-* The [SHA-1 hash](https://git-scm.com/book/en/v1/Getting-Started-Git-Basics#Git-Has-Integrity) of the GitHub commit of `targit` currently installed. `packageDescription("targit")$GithubSHA1` shows you this.
+* The [SHA-1 hash](https://git-scm.com/book/en/v1/Getting-Started-Git-Basics#Git-Has-Integrity) of the GitHub commit of `gittargets` currently installed. `packageDescription("gittargets")$GithubSHA1` shows you this.
