+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,7 +1,7 @@
 # Prework
 
 * [ ] I understand and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html).
-* [ ] I have already submitted a [discussion topic](https://github.com/wlandau/targit/discussions) or [issue](https://github.com/wlandau/targit/issues) to discuss my idea with the maintainer.
+* [ ] I have already submitted a [discussion topic](https://github.com/wlandau/gittargets/discussions) or [issue](https://github.com/wlandau/gittargets/issues) to discuss my idea with the maintainer.
 
 # Related GitHub issues and pull requests
 
