### 2021-12-22 6b3dada
Reorder tar_git_status() messages
