### 2021-12-27 #486 @mdneuzerling
Hello @adamhsparks @wlandau. I've made a start on this. Apologies for the delay --- I had to get Christmas out of the way first!

I've started by looking through the code structure to get a rough idea of things. [I'm making notes on a pull request](https://github.com/mdneuzerling/gittargets/pull/1), which I'll then compile into a formal review. This is a trick I've used before — check out the first commit of a repo into a separate branch, and then create a pull request onto it.
