+++ man/tar_git_snapshot.Rd
@@ -17,6 +17,7 @@ tar_git_snapshot(
   callr_arguments = targets::callr_args_default(callr_function, reporter),
   status = interactive(),
   force = FALSE,
+  pack_refs = TRUE,
   verbose = TRUE
 )
 }
@@ -52,7 +53,9 @@ to specify a path other than \verb{_targets/}. The path need not exist
 before the pipeline begins, and it need not end with "_targets",
 but it must be writeable.
 For optimal performance, choose a storage location
-with fast read/write access.}
+with fast read/write access.
+If the argument \code{NULL}, the setting is not modified.
+Use \code{\link[targets:tar_config_unset]{tar_config_unset()}} to delete a setting.}
 
 \item{stash_gitignore}{Logical of length 1, whether to temporarily
 stash the \code{.gitignore} file of the data store. See the
@@ -99,6 +102,11 @@ with \code{\link[=tar_git_status]{tar_git_status()}} and ask whether a snapshot 
 
 \item{force}{Logical of length 1. Force checkout the data branch
 of an existing data snapshot of the current code commit?}
+
+\item{pack_refs}{Logical of length 1, whether to run \verb{git pack-refs --all}
+in the data store after taking the snapshot. Packing references
+improves efficiency when the number of snapshots is large.
+Learn more at \url{https://git-scm.com/docs/git-pack-refs}.}
 
 \item{verbose}{Logical of length 1, whether to print R console messages
 confirming that a snapshot was created.}
