+++ tests/testthat/test-tar_git_log.R
@@ -1,11 +1,21 @@
 targets::tar_test("tar_git_log()", {
   skip_os_git()
   git_setup_init()
-  tar_git_snapshot(status = FALSE, verbose = FALSE)
+  tar_git_snapshot(status = FALSE, verbose = FALSE, message = "First snapshot")
   out <- tar_git_log()
-  expect_equal(dim(out), c(1L, 5L))
+  cols <- c(
+    "message_code",
+    "message_data",
+    "time_code",
+    "time_data",
+    "commit_code",
+    "commit_data"
+  )
+  expect_equal(sort(colnames(out)), sort(cols))
+  expect_equal(nrow(out), 1L)
   expect_false(any(is.na(out)))
   expect_equal(out$message_code, "First commit")
+  expect_equal(out$message_data, "First snapshot")
   for (field in c("time_code", "time_data")) {
     expect_s3_class(out[[field]], "POSIXct")
   }
