+++ NEWS.md
@@ -1,3 +1,3 @@
 # gittargets 0.0.0.9000
 
-* First release.
+* First version.
