### 2022-01-09 #486 @mdneuzerling
- I've previously used the author's `drake` and `targets` packages, and have had a small amount of personal communication regarding these. I've also published blog posts on these packages.

---

# Review Comments

The code in this package is to a very high standard. I have a few comments around documentation that I believe could improve a new user approaching this package for the first time. However, these are only recommendations, and I strongly recommend that this project be submitted to CRAN. I have some other comments that are nitpicks and candefinitely be ignored.

I created a very simple project to test the usage of this package. My code is minimal, and can be found at [mdneuzerling/gittargets-peerreview](https://github.com/mdneuzerling/gittargets-peerreview).

* I really appreciate the frequent use of assertions. This gives me confidence that the packages
* The documentation frequently repeats itself (a good thing). The author makes good use of the various inheritance tags in Roxygen.

## Recommendations

* I feel like the first paragraph of the README doesn't adequately capture the use case of `gittargets`. It also took me a while to understand. What I'm looking to see in the first few paragraphs is something that covers what the targets data store is, and what you want to have happen to it when you switch branches with git.
* In the project README or elsewhere, I recommend describing the intended workflow of the package. I settled on the below when checking out a different branch, which I think is right:
```r
# git checkout <branch>
tar_git_checkout() # to bring the data store up-to-date
# make changes
tar_make()
tar_git_snapshot()
```
* The assertions in `tar_git_log` allow for a value of `max = Inf`, which is a reasonable value that someone might use. This value is passed to `gert::git_log`, which will give a confusing error on `max = Inf`. Consider using `is.finite` or an equivalent assertion.
* I notice many tests are being skipped on Windows. I don't have ready access to a Windows machine. Can the author provide some confidence that this will work on Windows, or is the package intended to be Linux/Mac only? (I think we can safely disregard Solaris...)
* Regarding the vignette, the "Git data backend" is a local concept as far as I can tell. There's no use of a git _remote_ here. I think for many people, git and GitHub are entirely synonymous, so it would be good to distinguish between the local and remote.

## Nitpicks

* The `run` function from the `processx` package is in the NAMESPACE of `gittargets`. However in `R/utils_git.R` the author uses `processx::run`. Normally I wouldn't care, but `run` is a pretty commonly used function name, so I'd recommend keeping it out of the `NAMESPACE` if possible by using only `processx::run`.
* `tar_git_branch_snapshot` and `tar_git_commit_code` refer to a `"^code="` string; the meaning of this isn't entirely obvious (it's the prefix of the data branches). A comment in the code could clear this up.
* While it's not usually necessary to document internal functions, I think it would be a good idea to provide some Roxygen documentation for `tar_git_snapshot_menu` due to the large number of arguments it accepts. Recall that internal functions can be tagged with `@keywords internal` to prevent the help pages from being indexed.
* There are many SHA fingerprints in the diagrams which may make them harder to understand.
* The function name `tar_git_status_code` confuses me. "Status code" makes me think of something like a numeric HTTP status code or Unix exit status, but this returns a tibble.
* The function name `tar_git_status_targets` is not particularly clear to me.

## Session info

  <summary>Click to expand!</summary>

```
─ Session info  🤲🏾  🈁  🤯   ──────────────────────────────────────────────────────────────────────────
 hash: palms up together: medium-dark skin tone, Japanese “here” button, exploding head

 os       macOS Big Sur 11.3
 system   aarch64, darwin20
 ui       RStudio
 collate  en_AU.UTF-8
 ctype    en_AU.UTF-8
 tz       Australia/Melbourne
 rstudio  1.4.1717 Juliet Rose (desktop)
 pandoc   NA

─ Packages ────────────────────────────────────────────────────────────────────────────────────────────
 package     * version    date (UTC) lib source
 askpass       1.1        2019-01-13 [1] CRAN (R 4.1.0)
 backports     1.4.1      2021-12-13 [1] CRAN (R 4.1.1)
 base64url     1.4        2018-05-14 [1] CRAN (R 4.1.0)
 cachem        1.0.4      2021-02-13 [1] CRAN (R 4.1.0)
 callr         3.7.0      2021-04-20 [1] CRAN (R 4.1.0)
 cli           3.1.0      2021-10-27 [1] CRAN (R 4.1.1)
 codetools     0.2-18     2020-11-04 [1] CRAN (R 4.1.0)
 crayon        1.4.2      2021-10-29 [1] CRAN (R 4.1.1)
 credentials   1.3.2      2021-11-29 [1] CRAN (R 4.1.1)
 data.table    1.14.2     2021-09-27 [1] CRAN (R 4.1.1)
 desc          1.4.0      2021-09-28 [1] CRAN (R 4.1.1)
 devtools      2.4.0      2021-04-07 [1] CRAN (R 4.1.0)
 digest        0.6.29     2021-12-01 [1] CRAN (R 4.1.1)
 ellipsis      0.3.2      2021-04-29 [1] CRAN (R 4.1.0)
 fansi         0.5.0      2021-05-25 [1] CRAN (R 4.1.0)
 fastmap       1.1.0      2021-01-25 [1] CRAN (R 4.1.0)
 fs            1.5.0      2020-07-31 [1] CRAN (R 4.1.0)
 gert        * 1.5.0      2022-01-03 [1] CRAN (R 4.1.1)
 gittargets  * 0.0.0.9000 2022-01-08 [1] Github (wlandau/gittargets@6b3dada)
 glue          1.6.0      2021-12-17 [1] CRAN (R 4.1.1)
 igraph        1.2.11     2022-01-04 [1] CRAN (R 4.1.1)
 knitr         1.37       2021-12-16 [1] CRAN (R 4.1.1)
 lifecycle     1.0.1      2021-09-24 [1] CRAN (R 4.1.1)
 magrittr      2.0.1      2020-11-17 [1] CRAN (R 4.1.0)
 memoise       2.0.0      2021-01-26 [1] CRAN (R 4.1.0)
 openssl       1.4.6      2021-12-19 [1] CRAN (R 4.1.1)
 pillar        1.6.4      2021-10-18 [1] CRAN (R 4.1.1)
 pkgbuild      1.2.0      2020-12-15 [1] CRAN (R 4.1.0)
 pkgconfig     2.0.3      2019-09-22 [1] CRAN (R 4.1.0)
 pkgload       1.2.3      2021-10-13 [1] CRAN (R 4.1.1)
 prettyunits   1.1.1      2020-01-24 [1] CRAN (R 4.1.0)
 processx      3.5.2      2021-04-30 [1] CRAN (R 4.1.0)
 ps            1.6.0      2021-02-28 [1] CRAN (R 4.1.0)
 purrr         0.3.4      2020-04-17 [1] CRAN (R 4.1.0)
 R6            2.5.1      2021-08-19 [1] CRAN (R 4.1.1)
 remotes       2.3.0      2021-04-01 [1] CRAN (R 4.1.0)
 rlang         0.4.12     2021-10-18 [1] CRAN (R 4.1.1)
 rprojroot     2.0.2      2020-11-15 [1] CRAN (R 4.1.0)
 sessioninfo   1.2.1      2021-11-02 [1] CRAN (R 4.1.1)
 sys           3.4        2020-07-23 [1] CRAN (R 4.1.0)
 targets     * 0.10.0     2022-01-07 [1] CRAN (R 4.1.0)
 testthat      3.1.0      2021-10-04 [1] CRAN (R 4.1.1)
 tibble        3.1.6      2021-11-07 [1] CRAN (R 4.1.1)
 tidyselect    1.1.1      2021-04-30 [1] CRAN (R 4.1.0)
 usethis       2.0.1      2021-02-10 [1] CRAN (R 4.1.0)
 utf8          1.2.2      2021-07-24 [1] CRAN (R 4.1.0)
 uuid          1.0-3      2021-11-01 [1] CRAN (R 4.1.1)
 vctrs         0.3.8      2021-04-29 [1] CRAN (R 4.1.0)
 withr         2.4.3      2021-11-30 [1] CRAN (R 4.1.1)
 xfun          0.29       2021-12-14 [1] CRAN (R 4.1.1)
 yaml          2.2.1      2020-02-01 [1] CRAN (R 4.1.0)

 [1] /Library/Frameworks/R.framework/Versions/4.1-arm64/Resources/library
```
