### 2022-01-10 #486 @smwindecker
- I have used `targets` before, and have written twitter threads on the topic. I have communicated online with the author.
- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

---

Thank you for creating another useful addendum package to the `targets` family. I look forward to seeing this tool implemented in the suggested workflows for targets pipelines more broadly.

Statement of need:
* By the name I thought this package would be some sort of `targets` specific link to git itself (such as what `gert` is achieving in these examples).
* When I understood the purpose of the package, I was very excited about its usefulness, as snapshotting data outputs in a version controlled way was a question I asked myself using `targets` in the past, and even thought about git committing my _targets/ folder at one point!
* With that said, it took reading through most of the vignette and seeing the function outputs to really understand what the package was doing.
* The README and DESCRIPTION could use clearer statements, for example: "`gittargets` allows users to create version control snapshots of the data objects created in their pipeline." ... or something simple like that.

tar_git_status()
* I love this function! If I was still at all confused about how this package worked, the separate code git status and data git status made everything clear.
* Small note, at the first use (before a data snapshot has been made but after the git repo was initialised) we see “! Create the code repository with `gert::git_init()`.” under data git status header, but it shouldn’t be there as a code repo already exists.

Remotes
* Can a data snapshot be pushed to github, or some other place where it can be assigned, for example a "release" and downloaded? I imagine it could be useful in the scenario of a published/submitted manuscript to be able to push a single version of the snapshot associated with the final code push. Although it is naturally preferable that a reviewer recreate the pipeline from scratch for reproducibility purposes, for many slow pipelines that is unfeasible, and having a snapshot there for folks to examine and play with intermediate outputs would be fantastic as an option.

Vignette/instructions
* I found the steps easy to follow, but several things could be improved to make it easier to understand.
* Checking out a data snapshot defaults to checking out the latest in a given branch (yes?). I found myself more explicitly using the name of the data snapshot from the `tar_git_log` when checking it out to ensure I was selecting the appropriate one. Given this, it should be more obvious to users that data snapshots will, by default, be given the same name as the commit message of the code commit they are tied to, and that you can give a name to the snapshot.
* The idea of names of data snapshots is also a bit clouded by the "snapshot model" illustration, which shows the SHA instead of the commit name.
* The instructions could use another step to clarify what happens in a merge. Specifically that after merging a branch, the data snapshot does not get overwritten with the branch snapshot, and if desired, should be checked out at that point. Similarly, clarifying that deleting a branch does not delete its associated data commit.
* Is there a way to clean up outdated data snapshots? tar_git_remove_snapshot? Or is this unnecessary because the point is a version control record? I just imagine it could get too large very fast.
* The "snapshot model" illustration shows a linear model of working with `gittargets` that does not match the vignette steps, which delves into branches. I think the addition of a simple image, in the style of the one below that many of us are familiar with, could attempt to illustrate the separate pathways of the code and data in a branching workflow.

Internal functions
* When I write into my console `tar_git_` and hit tab, it shows me a range of functions that are not included in the exported functions (eg. `tar_git_commit()`) I found this confusing as I searched for the correct exported function to use, especially because these non exported but visible functions could not be searched with the `?` command (naturally, they're not exported!) It gave me pause that I was missing some functionalities.
* I _think_ these show up from the way the functions are documented as they sit in the same file as an exported function (I'm not at all sure this is the explanation).

Overall, my notes are quite minor and mostly related to improving user experience and uptake rather than real issues with the package. I love the idea and the execution is elegant, especially due to the function name parallels to the functions in `gert` (which I didn't know about before, so thank you!). I think developing a little more content about suggested best practices of how a `targets` user should integrate these steps into their workflow would make this contribution to the wider package family even more meaningful.

PS. the guitar in the hex is a very cute touch.
