+++ README.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 [![codecov](https://codecov.io/gh/wlandau/gittargets/branch/main/graph/badge.svg?token=3T5DlLwUVl)](https://app.codecov.io/gh/wlandau/gittargets)
 [![lint](https://github.com/wlandau/gittargets/workflows/lint/badge.svg)](https://github.com/wlandau/gittargets/actions?query=workflow%3Alint)
 
-Version control systems such as Git help researchers track changes and history in data science projects, and the [`targets`](https://docs.ropensci.org/targets/) package minimizes the computational cost of keeping the latest results reproducible and up to date. The `gittargets` package combines these two capabilities. The [`targets`](https://docs.ropensci.org/targets/) data store becomes a version control repository and stays synchronized with the Git repository of the source code. Users can switch commits and branches without invalidating the [`targets`](https://docs.ropensci.org/targets/) pipeline.
+Pipelines with the [`targets`](https://docs.ropensci.org/targets/) R package skip  steps that are up to already date. Although this behavior reduces the runtime of subsequent runs, it comes at the cost of overwriting previous results. Ordinarily, you would need to rerun the pipeline in order to recover the targets you had before. However, `gittargets` preserves historical output, creating version control snapshots of data store. Each data snapshot remembers the contemporaneous Git commit of the pipeline source code, so you can recover the right data when you navigate the Git history. In other words, `gittargets` makes it possible to switch commits or branches without invalidating the pipeline. You can simply check out the up-to-date targets from the past instead of taking the time to recompute them from scratch.
 
 ## Prerequisites
 
@@ -29,6 +29,7 @@ Version control systems such as Git help researchers track changes and history i
 1. [Data science workflow management best practices](https://rstats.wtf/index.html).
 1. [Git](https://git-scm.com), covered in [Happy Git and GitHub for the useR](https://happygitwithr.com).
 1. [`targets`](https://docs.ropensci.org/targets/), which has resources on the [documentation website](https://docs.ropensci.org/targets/#how-to-get-started).
+1. Familiarity with the [`_targets/` data store](https://books.ropensci.org/targets/files.html#internal-data-files).
 
 ## Installation
 
@@ -51,25 +52,66 @@ tar_git_ok()
 
 There are also backend-specific installation requirements and recommendations in the [package vignettes](https://wlandau.github.io/gittargets/articles/index.html).
 
-## How it works
+## Motivation
 
-![](./man/figures/workflow.png)
+Consider an example pipeline with source code in `_targets.R` and output in the [`_targets/` data store](https://books.ropensci.org/targets/files.html#internal-data-files). 
 
-### Development
+```{r, eval = FALSE}
+# _targets.R
+library(targets)
+list(
+  tar_target(data, airquality),
+  tar_target(model, lm(Ozone ~ Wind, data = data)) # Regress on wind speed.
+)
+```
 
-When you develop a reproducible analysis project, you write R code, run the pipeline, commit the code to a branch, and repeat. `gittargets` enters the picture at the end of each one of these development cycles. When the code files are up to date in the Git repository and the targets are up to date in the pipeline, `gittargets` can snapshot the [`targets` data store](https://books.ropensci.org/targets/files.html#internal-data-files) to a local version control repository of its own. Data snapshots map to code commits in a way that facilitates data recovery. This mapping may differ from backend to backend, and the [package vignettes](https://wlandau.github.io/gittargets/articles/index.html) explain the details.
+Suppose you run the pipeline and confirm that all targets are up to date.
 
-### Transitions
+```{r, eval = FALSE}
+tar_make()
+#> • start target data
+#> • built target data
+#> • start target model
+#> • built target model
+#> • end pipeline
+```
 
[…]
