+++ man/gittargets-package.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tar_package.R
-\docType{package}
-\name{gittargets-package}
-\alias{gittargets-package}
-\title{targets: Dynamic Function-Oriented Make-Like Declarative Pipelines for R}
-\description{
-Version control systems such as Git help researchers
-track changes and history in data science projects,
-and the \code{targets} package minimizes the computational cost of
-keeping the latest results reproducible and up to date.
-The \code{gittargets} package combines these two capabilities.
-The \code{targets} data store becomes a version control repository
-and stays synchronized with the Git repository of the source code.
-Users can switch commits and branches without
-invalidating the \code{targets} pipeline.
-}
-\concept{help}
