+++ README.md
@@ -9,16 +9,19 @@
 [![codecov](https://codecov.io/gh/wlandau/gittargets/branch/main/graph/badge.svg?token=3T5DlLwUVl)](https://app.codecov.io/gh/wlandau/gittargets)
 [![lint](https://github.com/wlandau/gittargets/workflows/lint/badge.svg)](https://github.com/wlandau/gittargets/actions?query=workflow%3Alint)
 
-Version control systems such as Git help researchers track changes and
-history in data science projects, and the
-[`targets`](https://docs.ropensci.org/targets/) package minimizes the
-computational cost of keeping the latest results reproducible and up to
-date. The `gittargets` package combines these two capabilities. The
-[`targets`](https://docs.ropensci.org/targets/) data store becomes a
-version control repository and stays synchronized with the Git
-repository of the source code. Users can switch commits and branches
-without invalidating the [`targets`](https://docs.ropensci.org/targets/)
-pipeline.
+Pipelines with the [`targets`](https://docs.ropensci.org/targets/) R
+package skip steps that are up to already date. Although this behavior
+reduces the runtime of subsequent runs, it comes at the cost of
+overwriting previous results. Ordinarily, you would need to rerun the
+pipeline in order to recover the targets you had before. However,
+`gittargets` preserves historical output, creating version control
+snapshots of data store. Each data snapshot remembers the
+contemporaneous Git commit of the pipeline source code, so you can
+recover the right data when you navigate the Git history. In other
+words, `gittargets` makes it possible to switch commits or branches
+without invalidating the pipeline. You can simply check out the
+up-to-date targets from the past instead of taking the time to recompute
+them from scratch.
 
 ## Prerequisites
 
@@ -32,6 +35,8 @@ pipeline.
 4.  [`targets`](https://docs.ropensci.org/targets/), which has resources
     on the [documentation
     website](https://docs.ropensci.org/targets/#how-to-get-started).
+5.  Familiarity with the [`_targets/` data
+    store](https://books.ropensci.org/targets/files.html#internal-data-files).
 
 ## Installation
 
@@ -65,45 +70,77 @@ There are also backend-specific installation requirements and
 recommendations in the [package
 vignettes](https://wlandau.github.io/gittargets/articles/index.html).
 
-## How it works
+## Motivation
 
-![](./man/figures/workflow.png)
-
-### Development
-
-When you develop a reproducible analysis project, you write R code, run
-the pipeline, commit the code to a branch, and repeat. `gittargets`
-enters the picture at the end of each one of these development cycles.
-When the code files are up to date in the Git repository and the targets
-are up to date in the pipeline, `gittargets` can snapshot the [`targets`
-data
-store](https://books.ropensci.org/targets/files.html#internal-data-files)
-to a local version control repository of its own. Data snapshots map to
-code commits in a way that facilitates data recovery. This mapping may
-differ from backend to backend, and the [package
-vignettes](https://wlandau.github.io/gittargets/articles/index.html)
-explain the details.
-
-### Transitions
-
-To reproducibly track alternative versions of your project, it is good
-practice to create multiple Git commits and Git branches for your R code
-during development. That way, you can transition among multiple versions
-of your code with
-[`git checkout`](https://git-scm.com/docs/git-checkout). But your code
-commits probably do not contain the [`targets` data
-store](https://books.ropensci.org/targets/files.html#internal-data-files)
-because it is too big to fit on [GitHub](https://github.com). If so,
-switching to a different branch will most likely invalidate the
-[`targets`](https://docs.ropensci.org/targets/) pipeline, causing you to
[…]

[…]
