+++ README.Rmd
@@ -29,7 +29,7 @@ Pipelines with the [`targets`](https://docs.ropensci.org/targets/) R package ski
 1. [Data science workflow management best practices](https://rstats.wtf/index.html).
 1. [Git](https://git-scm.com), covered in [Happy Git and GitHub for the useR](https://happygitwithr.com).
 1. [`targets`](https://docs.ropensci.org/targets/), which has resources on the [documentation website](https://docs.ropensci.org/targets/#how-to-get-started).
-1. Familiarity with the [`_targets/` data store](https://books.ropensci.org/targets/files.html#internal-data-files).
+1. Familiarity with the [`targets` data store](https://books.ropensci.org/targets/files.html#internal-data-files).
 
 ## Installation
 
@@ -54,7 +54,7 @@ There are also backend-specific installation requirements and recommendations in
 
 ## Motivation
 
-Consider an example pipeline with source code in `_targets.R` and output in the [`_targets/` data store](https://books.ropensci.org/targets/files.html#internal-data-files). 
+Consider an example pipeline with source code in `_targets.R` and output in the [data store](https://books.ropensci.org/targets/files.html#internal-data-files). 
 
 ```{r, eval = FALSE}
 # _targets.R
@@ -81,7 +81,7 @@ tar_outdated()
 #> character(0)
 ```
 
-It is good practice to track the source code in a [version control repository](https://git-scm.com) so you can revert to previous commits or branches. However, the [`_targets/` data store](https://books.ropensci.org/targets/files.html#internal-data-files) is usually too large to keep in the same repository as the code, which typically lives in a cloud platform like [GitHub](http://github.com) where space and bandwidth are pricey. So when you check out an old commit or branch, you revert the code, but not the data. In other words, your targets are out of sync and out of date.
+It is good practice to track the source code in a [version control repository](https://git-scm.com) so you can revert to previous commits or branches. However, the [data store](https://books.ropensci.org/targets/files.html#internal-data-files) is usually too large to keep in the same repository as the code, which typically lives in a cloud platform like [GitHub](http://github.com) where space and bandwidth are pricey. So when you check out an old commit or branch, you revert the code, but not the data. In other words, your targets are out of sync and out of date.
 
 ```{r, eval = FALSE}
 gert::git_branch_checkout(branch = "other-model")
@@ -103,11 +103,18 @@ tar_outdated()
 
 ## Usage
 
-With `gittargets`, you can keep your targets up to date even as you check out code from different commits or branches. For specific steps, please refer to the [package vignette](https://wlandau.github.io/gittargets/articles/index.html) of the `gittargets` backend you prefer. Currently, the only backend is [Git](https://git-scm.com)/[Git-LFS](https://git-lfs.github.com), but `gittargets` may support more backends in the future.
+With `gittargets`, you can keep your targets up to date even as you check out code from different commits or branches. The specific steps depend on the data backend you choose, and each supported backend has a [package vignette](https://wlandau.github.io/gittargets/articles/index.html) with a walkthrough. For example, the most important steps of the [Git data backend](https://wlandau.github.io/gittargets/articles/git.html) are as follows.
+
+1. Create the source code and run the pipeline at least once so the [data store](https://books.ropensci.org/targets/files.html#internal-data-files) exists.
+1. `tar_git_init()`: initialize a [Git](https://git-scm.com)/[Git-LFS](https://git-lfs.github.com) repository for the [data store](https://books.ropensci.org/targets/files.html#internal-data-files).
+1. Bring the pipeline up to date (e.g. with [`tar_make()`](https://docs.ropensci.org/targets/reference/tar_make.html)) and commit any changes to the source code.
+1. `tar_git_snapshot()`: create a data snapshot for the current code commit.
[…]
