### 2022-01-11 f12ee85
Assert finite max
