+++ tests/testthat/test-utils_assert.R
@@ -62,3 +62,9 @@ targets::tar_test("tar_git_assert_repo_data() no branch", {
   branch <- tar_git_branch_snapshot(tar_git_log()$commit_code)
   expect_silent(tar_git_assert_snapshot(branch = branch, store = store))
 })
+
+tar_test("tar_assert_finite", {
+  expect_silent(tar_assert_finite(1))
+  expect_silent(tar_assert_finite(c(1, 2)))
+  expect_error(tar_assert_finite(c(1, Inf)), class = "tar_condition_validate")
+})
