### 2022-01-11 5acd4ec
Only skip on Windows if also on CI
