+++ README.md
@@ -145,9 +145,8 @@ follows.
     [data
     store](https://books.ropensci.org/targets/files.html#internal-data-files)
     exists.
-2.  `tar_git_init()`: initialize a
-    [Git](https://git-scm.com)/[Git-LFS](https://git-lfs.github.com)
-    repository for the [data
+2.  `tar_git_init()`: initialize a [Git](https://git-scm.com)/[Git
+    LFS](https://git-lfs.github.com) repository for the [data
     store](https://books.ropensci.org/targets/files.html#internal-data-files).
 3.  Bring the pipeline up to date (e.g. with
     [`tar_make()`](https://docs.ropensci.org/targets/reference/tar_make.html))
@@ -175,7 +174,7 @@ The first data versioning system in `gittargets` uses
 [Git](https://git-scm.com), which is designed for source code and may
 not scale to enormous amounts of compressed data. Future releases of
 `gittargets` may explore alternative data backends more powerful than
-[Git-LFS](https://git-lfs.github.com).
+[Git LFS](https://git-lfs.github.com).
 
 ## Alternatives
 
@@ -236,4 +235,4 @@ citation("gittargets")
 
 [1] `gert` does not have these requirements, but `gittargets` does not
 exclusively rely on `gert` because `libgit2` does not automatically work
-with `git-lfs`.
+with [Git LFS](https://git-lfs.github.com).
