+++ README.Rmd
@@ -40,7 +40,7 @@ install.packages("remotes")
 remotes::install_github("wlandau/gittargets")
 ```
 
-You will also need command line Git, available at <https://git-scm.com/downloads>.^[`gert` does not have these requirements, but `gittargets` does not exclusively rely on `gert` because `libgit2` does not automatically work with `git-lfs`.] Please make sure Git is reachable from your system path environment variables. To control which Git executable `gittargets` uses, you may set the `TAR_GIT` environment variable with `usethis::edit_r_environ()` or `Sys.setenv()`. You will also need to configure your user name and user email at the global level using the instructions at <https://git-scm.com/book/en/v2/Getting-Started-First-Time-Git-Setup> (or `gert::git_config_global_set()`). Run `tar_git_ok()` to check installation and configuration.
+You will also need command line Git, available at <https://git-scm.com/downloads>.^[`gert` does not have these requirements, but `gittargets` does not exclusively rely on `gert` because `libgit2` does not automatically work with [Git LFS](https://git-lfs.github.com).] Please make sure Git is reachable from your system path environment variables. To control which Git executable `gittargets` uses, you may set the `TAR_GIT` environment variable with `usethis::edit_r_environ()` or `Sys.setenv()`. You will also need to configure your user name and user email at the global level using the instructions at <https://git-scm.com/book/en/v2/Getting-Started-First-Time-Git-Setup> (or `gert::git_config_global_set()`). Run `tar_git_ok()` to check installation and configuration.
 
 ```{r, eval = FALSE}
 tar_git_ok()
@@ -106,7 +106,7 @@ tar_outdated()
 With `gittargets`, you can keep your targets up to date even as you check out code from different commits or branches. The specific steps depend on the data backend you choose, and each supported backend has a [package vignette](https://wlandau.github.io/gittargets/articles/index.html) with a walkthrough. For example, the most important steps of the [Git data backend](https://wlandau.github.io/gittargets/articles/git.html) are as follows.
 
 1. Create the source code and run the pipeline at least once so the [data store](https://books.ropensci.org/targets/files.html#internal-data-files) exists.
-1. `tar_git_init()`: initialize a [Git](https://git-scm.com)/[Git-LFS](https://git-lfs.github.com) repository for the [data store](https://books.ropensci.org/targets/files.html#internal-data-files).
+1. `tar_git_init()`: initialize a [Git](https://git-scm.com)/[Git LFS](https://git-lfs.github.com) repository for the [data store](https://books.ropensci.org/targets/files.html#internal-data-files).
 1. Bring the pipeline up to date (e.g. with [`tar_make()`](https://docs.ropensci.org/targets/reference/tar_make.html)) and commit any changes to the source code.
 1. `tar_git_snapshot()`: create a data snapshot for the current code commit.
 1. Develop the pipeline. Creating new code commits and code branches early and often, and create data snapshots at key strategic milestones.
@@ -118,7 +118,7 @@ With `gittargets`, you can keep your targets up to date even as you check out co
 
 ## Future directions
 
-The first data versioning system in `gittargets` uses [Git](https://git-scm.com), which is designed for source code and may not scale to enormous amounts of compressed data. Future releases of `gittargets` may explore alternative data backends more powerful than [Git-LFS](https://git-lfs.github.com).
+The first data versioning system in `gittargets` uses [Git](https://git-scm.com), which is designed for source code and may not scale to enormous amounts of compressed data. Future releases of `gittargets` may explore alternative data backends more powerful than [Git LFS](https://git-lfs.github.com).
 
 ## Alternatives
 
