### 2022-01-11 541168f
tar_git_snapshot_menu() docstring
