### 2022-01-11 wlandau/gittargets#9 @wlandau
Function to remove snapshots
## Prework

* [x] Read and agree to the [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html) and [contributing guidelines](https://github.com/wlandau/gittargets/blob/main/CONTRIBUTING.md).
* [x] If there is [already a relevant issue](https://github.com/wlandau/gittargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
* [x] New features take time and effort to create, and they take even more effort to maintain. So if the purpose of the feature is to resolve a struggle you are encountering personally, please consider first posting a "trouble" or "other" issue so we can discuss your use case and search for existing solutions first.
* [x] Format your code according to the [tidyverse style guide](https://style.tidyverse.org/).

## Proposal

Suggested by @smwindecker in her rOpenSci review. Plan:

1. Make all git snapshots orphan branches. History will no longer be shared among branches, but there is no loss of efficiency for fully up-to-date files. Efficiency from diffs may be lost, but diffs in target storage may be impossible anyway, I will need to check.
2. Remove commits with no ref: https://stackoverflow.com/questions/3765234/listing-and-deleting-git-commits-that-are-under-no-branch-dangling
