### 2022-01-12 37a20a7
Redo snapshot model figure
