### 2022-01-12 befd2ee
Improve tar_git_snapshot() docstring
