### 2022-01-12 40ae234
Clarify default code commit in tar_git_checkout()
