+++ R/tar_git_checkout.R
@@ -8,8 +8,8 @@
 #'   or other reference in the code repository
 #'   that points to a code commit. (You can also identify the code
 #'   commit by supplying a data branch of the form `code=<SHA1>`.)
-#'   Defaults to `"HEAD"`, which points to the latest code commit
-#'   in the current branch.
+#'   Defaults to `"HEAD"`, which points to the currently
+#'   checked out code commit.
 #'
 #'   Once the desired code commit is identified,
 #'   `tar_git_snapshot()` checks out the latest corresponding data snapshot.
