### 2022-01-12 wlandau/gittargets#9 @wlandau
From https://git-scm.com/book/id/v2/Kostumisasi-Git-Git-Attributes, it looks like file diffs rely on filters. Gittargets already uses the lfs filter, which seems to preclude diffs, so orphan branches are probably fine.
