### 2022-01-12 wlandau/gittargets#9 @wlandau
On second thought, orphan branches lose valuable information about the audit trail of which targets changed between code commits. From the perspective of tracking history, I think this is more valuable than the ability to delete data snapshots. Might revisit this issue at some point.
