### 2022-01-12 #486 @wlandau
# Response to @mdneuzerling

> I feel like the first paragraph of the README doesn't adequately capture the use case of gittargets. What I'm looking to see in the first few paragraphs is something that covers what the targets data store is, and what you want to have happen to it when you switch branches with git.

I expanded the README in https://github.com/wlandau/gittargets/commit/11943f11b1f719e0f815f10e31f2290c8b6c514c to describe those assumptions.

> In the project README or elsewhere, I recommend describing the intended workflow of the package. I settled on the below when checking out a different branch, which I think is right:

https://github.com/wlandau/gittargets/commit/ddc1e59f9560131e2aef971d1dfe122568762930 replaces `man/figures/workflow.png` with a simpler description of the workflow in the Usage section of the README and at the top of the `git.Rmd` vignette.

> The assertions in tar_git_log allow for a value of max = Inf, which is a reasonable value that someone might use. This value is passed to gert::git_log, which will give a confusing error on max = Inf. Consider using is.finite or an equivalent assertion.

Added an assertion in https://github.com/wlandau/gittargets/commit/f12ee85401f607b582bd321283b6ecea21e95101.

> I notice many tests are being skipped on Windows. I don't have ready access to a Windows machine. Can the author provide some confidence that this will work on Windows, or is the package intended to be Linux/Mac only? (I think we can safely disregard Solaris...)

When I unskip those tests on my local Windows machine, they all pass. But I have not been able to get them to work on GitHub Actions because the Windows runner claims the user name and email are not set, despite multiple attempted workarounds. As of https://github.com/wlandau/gittargets/commit/5acd4ecf648986b327d056abe4ed869aaf23b465 these tests are only skipped on CI Windows, not local Windows.

I would have preferred to use `gert` for all Git-related functionality because it installs well and just works on all platforms without any config. But `gert` relies on `libgit2`, which still silently fails with Git LFS unless the user makes some custom modifications: https://github.com/git-lfs/git-lfs/issues/375, https://github.com/libgit2/libgit2/issues/1473. So now `gittargets` requires a full Git client for operations that modify local Git repos. For reading stuff from Git, `gittargets` still uses `gert` because it gets the output into R quite nicely.

> Regarding the vignette, the "Git data backend" is a local concept as far as I can tell. There's no use of a git remote here. I think for many people, git and GitHub are entirely synonymous, so it would be good to distinguish between the local and remote.

https://github.com/wlandau/gittargets/commit/bfe062009873cb20efe37700ee29d05cbbd9c610 adds a section on remotes and emphasizes that `gittargets` uses local Git only.

> The run function from the processx package is in the NAMESPACE of gittargets. However in R/utils_git.R the author uses processx::run. Normally I wouldn't care, but run is a pretty commonly used function name, so I'd recommend keeping it out of the NAMESPACE if possible by using only processx::run.

I see what you mean, but importing `process::run()` in the `NAMESPACE` file only has consequences inside the `gittargets` private namespace, which does not have a conflicting internal function called "run". And it should not conflict with external code unless I re-export it. And think it is good practice to have `NAMESPACE` entries for `DESCRIPTION` `Imports:` packages (even with namespaced calls in the code, which I also do for the sake of defensive programming.)

> tar_git_branch_snapshot and tar_git_commit_code refer to a "^code=" string; the meaning of this isn't entirely obvious (it's the prefix of the data branches). A comment in the code could clear this up.

Added comments in https://github.com/wlandau/gittargets/commit/5b7b1425fe5cb916c67dfd2283d14774f7c4b59a.

> While it's not usually necessary to document internal functions, I think it would be a good idea to provide some Roxygen documentation for tar_git_snapshot_menu due to the large number of arguments it accepts. Recall that internal functions can be tagged with @keywords internal to prevent the help pages from being indexed.

Added a docstring in https://github.com/wlandau/gittargets/commit/541168f1c405dda019a4c05e0025a210f92a5327.

> There are many SHA fingerprints in the diagrams which may make them harder to understand.

It may look like a lot to take in at first glance, but SHAs are important for understanding what is going on because of how the "code=<SHA1>" data branches match up with code commits. I kept them but simplified the figure in https://github.com/wlandau/gittargets/commit/37a20a75d7149bed889071570a308c4d56972aa9.

> The function name tar_git_status_code confuses me. "Status code" makes me think of something like a numeric HTTP status code or Unix exit status, but this returns a tibble.

I get what you mean. I guess I could change it to a name like `tar_git_status_code_repo()`, but I am not 100% sold at the moment because it would interrupt the internal consistency of the function name convention.
