### 2022-01-12 #486 @wlandau
# Response to @smwindecker

> With that said, it took reading through most of the vignette and seeing the function outputs to really understand what the package was doing.
> The README and DESCRIPTION could use clearer statements, for example: "gittargets allows users to create version control snapshots of the data objects created in their pipeline." ... or something simple like that.

I rewrote the README and DESCRIPTION in https://github.com/wlandau/gittargets/commit/11943f11b1f719e0f815f10e31f2290c8b6c514c, hopefully they are clearer now.

> Small note, at the first use (before a data snapshot has been made but after the git repo was initialised) we see “! Create the code repository with gert::git_init().” under data git status header, but it shouldn’t be there as a code repo already exists.

https://github.com/wlandau/gittargets/commit/ef8c0e07f87802ba3c6580b27b10377c514b7a83 improves the `tar_git_status()` messages.

> Can a data snapshot be pushed to github, or some other place where it can be assigned, for example a "release" and downloaded? I imagine it could be useful in the scenario of a published/submitted manuscript to be able to push a single version of the snapshot associated with the final code push. Although it is naturally preferable that a reviewer recreate the pipeline from scratch for reproducibility purposes, for many slow pipelines that is unfeasible, and having a snapshot there for folks to examine and play with intermediate outputs would be fantastic as an option.

I think something like this should be possible if you create a Git-LFS-backed repo on GitHub or similar. Data snapshots are just branches, and you could either selectively push them to the remote repo, or you could push them all with [`git push --all; git push --tags`](https://stackoverflow.com/questions/6865302/push-local-git-repo-to-new-remote-including-all-branches-and-tags). Then a code release could refer to a branch and commit in the remote data repo. [`gittargets` itself tries not to go beyond local Git](https://github.com/wlandau/gittargets/blob/ef8c0e07f87802ba3c6580b27b10377c514b7a83/vignettes/git.Rmd#L33). But ultimately, this is the user's responsibility to manage outside `gittargets`.

> Checking out a data snapshot defaults to checking out the latest in a given branch (yes?). I found myself more explicitly using the name of the data snapshot from the tar_git_log when checking it out to ensure I was selecting the appropriate one. Given this, it should be more obvious to users that data snapshots will, by default, be given the same name as the commit message of the code commit they are tied to, and that you can give a name to the snapshot.

As of https://github.com/wlandau/gittargets/commit/40ae23469486196628a426b7210fdf15beb8697c, the docstring of `tar_git_checkout()` is clearer. By default, `tar_git_checkout()` checks out the snapshot from the currently checked out code commit ("HEAD"), which is not necessarily the latest in the current branch (my mistake in the previous docstring). In other words, if you do a `git reset --hard old_commit` or a `git checkout old_branch` and then call `tar_git_checkout()`, then you could get a very old (but still correct) data snapshot without supplying the `ref` argument. The idea is to let the user focus on the code repo and trust `tar_git_snapshot()` to automatically find the right data. So ideally, the user should almost never need the `ref` argument. Does that make sense?

> The idea of names of data snapshots is also a bit clouded by the "snapshot model" illustration, which shows the SHA instead of the commit name.

I added commit messages to the new figure in https://github.com/wlandau/gittargets/commit/37a20a75d7149bed889071570a308c4d56972aa9.

> The instructions could use another step to clarify what happens in a merge. Specifically that after merging a branch, the data snapshot does not get overwritten with the branch snapshot, and if desired, should be checked out at that point. Similarly, clarifying that deleting a branch does not delete its associated data commit.

Yes, ordinary Git operations on the code repo do not change the data repo. https://github.com/wlandau/gittargets/commit/b26a7f5e64be8a728b8e0320ee94cc32f481a0dc adds a small section to the vignette on merges specifically.

> Is there a way to clean up outdated data snapshots? tar_git_remove_snapshot? Or is this unnecessary because the point is a version control record? I just imagine it could get too large very fast.

I would have liked to implement this, but unfortunately it seems like an uphill battle against the intended behavior and design of Git. Because `gittargets` creates a large number of single-commit branches, the usual squash/rebase approach for deleting commits appears to breaks down. (I suppose `git merge --squash` comes close, but then that creates a new squash commit, and we would need to put that in a different data branch somehow, which seems messy.) An alternative I explored in https://github.com/wlandau/gittargets/issues/9 is to create an orphan branch for each data snapshot, but then we would lose the automatic audit trail of which targets changed between data snapshots, which I would prefer to keep for the sake of reproducibility.

In the `gittargets` documentation, I do recommend that snapshots only be created at important strategic milestones. This is the opposite of the usual advice for code version control (i.e. to commit early and often). The code and the data have such different needs when it comes to version control, and this fact was actually a major motivation for `gittargets`, which operates on entirely different Git repos for the code vs the data.

> The "snapshot model" illustration shows a linear model of working with gittargets that does not match the vignette steps, which delves into branches. I think the addition of a simple image, in the style of the one below that many of us are familiar with, could attempt to illustrate the separate pathways of the code and data in a branching workflow.

There is a new figure in https://github.com/wlandau/gittargets/commit/37a20a75d7149bed889071570a308c4d56972aa9 which more closely aligns with the vignette and the style of https://git-scm.com/book/en/v2/Git-Branching-Branches-in-a-Nutshell.

> When I write into my console tar_git_ and hit tab, it shows me a range of functions that are not included in the exported functions (eg. tar_git_commit()) I found this confusing as I searched for the correct exported function to use, especially because these non exported but visible functions could not be searched with the ? command (naturally, they're not exported!) It gave me pause that I was missing some functionalities.
> I think these show up from the way the functions are documented as they sit in the same file as an exported function (I'm not at all sure this is the explanation).

This happens to me if I load `gittargets` with `pkgload::load_all()`, but not if I restart my R session and load it with `library(gittargets)`. Does that help?
