### 2022-01-12 #486 @mdneuzerling
I think @wlandau's responses to my recommendations are more than adequate. In particular, the improvements to the README are superb. I don't have anything further to say about this package, other than that I look forward to seeing it on CRAN and using it in my own projects. Thank you so much for your contributions, @wlandau.
