### 2022-01-12 #486 @smwindecker
Thanks @wlandau for your responses.

The changes and explanations made much clearer the concept of the data snapshots as branches themselves. The relevance of this to understanding the snapshot 'ref's and how to call them is now clear. I appreciate the new figure and especially the accompanying explanation. And yes I had a local copy of the whole package and that's why I was seeing all the functions!

I am satisfied my review has been fully addressed. Look forward to sharing this with others.
