### 2022-01-12 #486 @ropensci-review-bot
Approved! Thanks @wlandau for submitting and @smwindecker, @mdneuzerling for your reviews! :grin:
