### 2022-01-12 #486 @adamhsparks
Well done, @wlandau! Thanks for this new package submission to rOpenSci. I'm guessing you know the drill now. If you need assistance with anything, please let me know.
