### 2022-01-12 #486 @wlandau
Amazing! Thanks @adamhsparks for facilitating the process and thanks @smwindecker and @mdneuzerling for the reviews! `gittargets` is in much better shape now, especially regarding documentation.

@adamhsparks, would you be willing to grant me admin permissions at https://github.com/ropensci/gittargets? The rOpenSci transfer process still removes that on my end.
