### 2022-01-12 #486 @wlandau
Ah, sorry I didn't see this:

> After transfer write a comment @ropensci-review-bot finalize transfer of <package-name> where <package-name> is the repo/package name. This will give you admin access back.
