### 2022-01-18 #486 @wlandau
Very minor: on my end, it looks like the rOpenSci README badge still says "Under Review". Are these badges still supported?

![Screen Shot 2022-01-18 at 3 03 39 PM](https://user-images.githubusercontent.com/1580860/150010539-c93478b7-8c1c-4954-bc8b-74b2fb634813.png)
