### 2022-01-18 #486 @mpadge
Sorry @wlandau, our badge server went to sleep a month ago without anybody realising - I've re-enabled it, so it should update soon.
